requirements.txt
apt-xapian-index==0.49
asgiref==3.3.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defusedxml==0.7.0rc2
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.4
django-braces==1.14.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-model-utils==4.1.1
django-paypal==1.0.0
django-star-ratings==0.9.1
django-tables2==2.3.4
django-taggit==1.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
entrypoints==0.3
gpg===1.13.1-unknown
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.3
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
psycopg2-binary==2.8.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
swapper==1.1.2.post1
systemd-python==234
ttystatus==0.38
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
